requirements.txt
protobuf==4.24.4
streamlit==1.27.2
google-generativeai==0.2.1
